Make metering API as lib, avoid the complex inherit relationship with ipsec
