+++ neutron_vpnaas/services/vpn/device_drivers/ovn_ipsec.py
@@ -15,6 +15,7 @@
 
 import netaddr
 
+from oslo_config import cfg
 from oslo_log import log as logging
 
 from neutron.agent.common import utils as agent_common_utils
@@ -25,8 +26,8 @@ from neutron_vpnaas.services.vpn.common import topics
 from neutron_vpnaas._i18n import _LE, _LI
 
 from neutron_vpnaas.services.vpn.device_drivers import ipsec
+from neutron_vpnaas.services.vpn.device_drivers import meter_manager
 from neutron_vpnaas.services.vpn.device_drivers import strongswan_ipsec
-from neutron_vpnaas.services.vpn.device_drivers import meter_ipsec
 
 OVN_NS_PREFIX = 'qvpn-'
 
@@ -34,6 +35,19 @@ port_prefix = {'external': 'vg', 'internal': 'vr'}
 
 LOG = logging.getLogger(__name__)
 
+meter_opts = [
+    cfg.BoolOpt('vpn_meter_enable',
+                default=False,
+                help=_('Enable flag for VPN metering function')),
+    cfg.IntOpt('vpn_measure_interval',
+               default=30,
+               help=_('The interval between two metering measures')),
+    cfg.IntOpt('vpn_report_interval',
+               default=300,
+               help=_('The interval between two metering reports'))
+]
+
+cfg.CONF.register_opts(meter_opts, 'meter')
 
 class DeviceManager(object):
     """Device Manager for ports in qvpn-xx namespace.
@@ -184,18 +198,12 @@ class NamespaceManager(object):
             LOG.exception(msg, name)
 
 
-class OvnOpenSwanProcess(ipsec.OpenSwanProcess, meter_ipsec.MeterProcess):
-    
-    def __init__(self, conf, process_id, vpnservice, namespace):
-        super(OvnOpenSwanProcess).__init__(conf, process_id, vpnservice, namespace)
-	meter_ipsec.MeterProcess.__init__(self, conf, process_id, vpnservice, namespaces)
+class OvnOpenSwanProcess(ipsec.OpenSwanProcess):
+    pass
 
 
-class OvnStrongSwanProcess(strongswan_ipsec.StrongSwanProcess, meter_ipsec.MeterProcess):
-    
-    def __init__(self, conf, process_id, vpnservice, namespace):
-        super(OvnStrongSwanProcess, self).__init__(conf, process_id, vpnservice, namespace)
-	meter_ipsec.MeterProcess.__init__(self, conf, process_id, vpnservice, namespace)
+class OvnStrongSwanProcess(strongswan_ipsec.StrongSwanProcess):
+    pass
 
 
 class IPsecOvnDriverApi(ipsec.IPsecVpnDriverApi):
@@ -219,7 +227,7 @@ class IPsecOvnDriverApi(ipsec.IPsecVpnDriverApi):
                           ptype=ptype, router_id=router_id, host=host)
 
 
-class OvnSwanDriver(meter_ipsec.MeterIPsecDriver, ipsec.IPsecDriver):
+class OvnSwanDriver(ipsec.IPsecDriver):
 
     def __init__(self, vpn_service, host):
         self.nsmgr = NamespaceManager()
@@ -228,6 +236,10 @@ class OvnSwanDriver(meter_ipsec.MeterIPsecDriver, ipsec.IPsecDriver):
         self.agent_rpc = IPsecOvnDriverApi(topics.IPSEC_DRIVER_TOPIC)
         self.devmgr = DeviceManager(self.conf, self.host,
                         self.agent_rpc, self.context)
+
+        if self.conf.meter.vpn_meter_enable:
+            self.metermgr = meter_manager.MeterManager(self.conf, self.host,
+                                                       self.context)
 
     def prepare_namespace(self, context, **kwargs):
         router = kwargs.get('router', None)
@@ -271,9 +283,20 @@ class OvnSwanDriver(meter_ipsec.MeterIPsecDriver, ipsec.IPsecDriver):
 
     def destroy_process(self, process_id):
         LOG.info(_LI('process %s is destroyed') % process_id)
+        namespace = self.devmgr.get_namespace_name(process_id)
+
+        if self.conf.meter.vpn_meter_enable:
+            self.metermgr.add_metering_infos()
+            if process_id in self.processes:
+                process = self.processes[process_id]
+                vpnservice = process.vpnservice
+                tenant_id = vpnservice['project_id']
+                self.metermgr.clean_tenant_conn_mapping(vpnservice)
+                self.metermgr.update_metering_rule(vpnservice, namespace, [], self.metermgr.remove_metering_rule)
+                del self.metermgr.namespaces[tenant_id]
+
         super(OvnSwanDriver, self).destroy_process(process_id)
 
-        namespace = self.devmgr.get_namespace_name(process_id)
         self.devmgr.del_static_routes(namespace)
 
     def create_router(self, router):
@@ -308,10 +331,16 @@ class OvnSwanDriver(meter_ipsec.MeterIPsecDriver, ipsec.IPsecDriver):
                                               vpnservice=vpnservice)
                 self._update_route(vpnservice)
                 self._update_nat(vpnservice, self.add_nat_rule)
-		if self.conf.meter.vpn_meter_enable:
-                    conn_del = process.conn_id_check(vpnservice)
-                    self._sync_tenant_conn_mapping(vpnservice, conn_del)
-                    self._update_metering_rule(vpnservice, conn_del, self.add_metering_rule)
+
+                if self.conf.meter.vpn_meter_enable:
+                    router_id = vpnservice['router_id']
+                    ns = self.get_namespace(router_id)
+                    tenant_id = vpnservice['project_id']
+                    self.metermgr.namespaces[tenant_id] = ns
+                    conn_del = self.metermgr.conn_id_check(process.etc_dir, vpnservice)
+                    self.metermgr.sync_tenant_conn_mapping(vpnservice, conn_del)
+                    self.metermgr.update_metering_rule(vpnservice, ns, conn_del, self.metermgr.add_metering_rule)
+
                 router = self.routers.get(vpnservice['router_id'])
                 if not router:
                     continue
